requirements.txt
beautifulsoup4=="4.12.3"
streamlit=="1.37.0"
pinecone-client=="5.0.1"
openai=="1.38.0"